Separated header into its own config file added revalidation
