+++ src/MainNav/hooks/revalidateNav.ts
@@ -1,12 +1,12 @@
 import type { GlobalAfterChangeHook } from 'payload';
 import { revalidateTag } from 'next/cache';
 
-export const revalidateHeader: GlobalAfterChangeHook = ({
+export const revalidateNav: GlobalAfterChangeHook = ({
   doc,
   req: { payload, context },
 }) => {
   if (!context.disableRevalidate) {
-    payload.logger.info(`Revalidating header`);
+    payload.logger.info(`Revalidating navigation`);
 
     revalidateTag('global_header');
   }
